- compile fixes: ser & tm + tcp compiles on linux, freebsd, openbsd & netbsd

+++ sip_router/parser/msg_parser.c
@@ -1,5 +1,5 @@
 /*
- * $Id: msg_parser.c,v 1.26 2003/01/29 19:24:10 jiri Exp $
+ * $Id: msg_parser.c,v 1.27 2003/02/20 18:13:22 andrei Exp $
  *
  * sip msg. header proxy parser 
  *
@@ -173,9 +173,10 @@ char* get_hdr_field(char* buf, char* end, struct hdr_field* hdr)
 				LOG(L_ERR, "ERROR:get_hdr_field: bad content_length header\n");
 				goto error;
 			}
-			hdr->parsed=(void*)integer;
+			hdr->parsed=(void*)(long)integer;
 			hdr->body.len=tmp-hdr->body.s;
-			DBG("DEBUG: get_hdr_body : content_length=%d\n",(int)hdr->parsed);
+			DBG("DEBUG: get_hdr_body : content_length=%d\n",
+					(int)(long)hdr->parsed);
 			break;
 		case HDR_CONTENTTYPE:
 		case HDR_FROM:
