+++ sip_router/parser/parse_content.h
@@ -1,5 +1,5 @@
 /*
- * $Id: parse_content.h,v 1.2 2003/01/24 16:44:43 bogdan Exp $
+ * $Id: parse_content.h,v 1.3 2003/02/20 18:13:22 andrei Exp $
  *
  *
  * Copyright (C) 2001-2003 Fhg Fokus
@@ -45,13 +45,13 @@
 /*
  * returns the content-length value of a sip_msg as an integer
  */
-#define get_content_length(_msg_)   ((int)((_msg_)->content_length->parsed))
+#define get_content_length(_msg_)   ((long)((_msg_)->content_length->parsed))
 
 
 /*
  * returns the content-type value of a sip_msg as an integer
  */
-#define get_content_type(_msg_)   ((int)((_msg_)->content_type->parsed))
+#define get_content_type(_msg_)   ((long)((_msg_)->content_type->parsed))
